requirements.txt
aiosqlite==0.17.0
alembic==1.13.1
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.3
asyncpg==0.29.0
certifi==2024.6.2
cffi==1.16.0
click==8.1.7
fastapi==0.108.0
greenlet==3.0.3
h11==0.14.0
idna==3.7
Mako==1.3.5
MarkupSafe==2.1.5
minio==7.2.7
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.4
pydantic_core==2.18.4
python-dotenv==1.0.0
python-multipart==0.0.9
redis==5.0.3
sniffio==1.3.1
SQLAlchemy==2.0.25
starlette==0.32.0.post1
typing_extensions==4.12.2
urllib3==2.2.1
uvicorn==0.25.0
